Function

+++ app/Http/Controllers/RideController.php
@@ -6,10 +6,14 @@ use App\Models\Car;
 use App\Models\Ride;
 use App\Models\User;
 use App\Models\Place;
+use App\Mail\denyMail;
 use App\Models\Picture;
 use Illuminate\Http\Request;
+use App\Mail\confimartionMail;
 use Illuminate\Support\Carbon;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Mail;
 use Illuminate\Support\Facades\Crypt;
 
 class RideController extends Controller
@@ -63,11 +67,24 @@ class RideController extends Controller
     public function show(Ride $ride)
     {
         //dd($ride->users); 
+        $pass = DB::table('ride_user')
+        ->where('user_id', Auth::user()->id)
+        ->where('ride_id', $ride->id)
+        ->count();
+
+        $taken = DB::table('ride_user')
+        ->where('ride_id', $ride->id)
+        ->where('approved_u', true)
+        ->count();
+
+        $leftSits = $ride->passengers_t - $taken;
+
+        //dd($pass);
         $pic = Picture::where('cars_id', $ride->cars_id)
         ->where('type_p', 4)
         ->first();
         $placa = Crypt::decryptString($ride->cars->placas_c);
-        return view('rides.showRide', compact('ride', 'placa', 'pic'));
+        return view('rides.showRide', compact('ride', 'placa', 'pic', 'pass', 'leftSits'));
     }
 
     /**
@@ -118,37 +135,56 @@ class RideController extends Controller
         return view('rides.myRides', compact('rides'));
     }
 
+    public function ridesImIn(){
+        $user = Auth::user();
+        return view('rides.ridesImIn', compact('user'));
+    }
+
     public function seeStops(Ride $ride){
         $places = Place::all();
         return view('rides.addStops', compact('ride', 'places'));
     }
 
     public function manageStops(Request $request, Ride $ride){
-        //dd($request);
         $ride->stops()->sync($request->place_id);
         return redirect()->route('ride.show', $ride);
     }
 
     public function requestStop(Request $request, Ride $ride){
-        //dd($request);
         $ride->users()->attach(Auth::user()->id, ['places_id'=>$request->places_id, 'approved_u'=>false]);
-        //$ride->stops()->sync($request->place_id);
         return redirect()->route('ride.show', $ride);
     }
 
     public function approveStop(Ride $ride, Place $place, User $user){
         //dd($ride);
-        $ride->users()->detach($user->id);
-        $ride->users()->attach($user->id, ['places_id'=>$place->id, 'approved_u'=>true]);
-        //CORREO
+        $taken = DB::table('ride_user')
+        ->where('ride_id', $ride->id)
+        ->where('approved_u', true)
+        ->count();
+
+        if($taken < $ride->passengers_t){
+            $ride->users()->detach($user->id);
+            $ride->users()->attach($user->id, ['places_id'=>$place->id, 'approved_u'=>true]);
+            $this->sendMail($user, $ride, $place);
+        }
         return redirect()->route('ride.show', $ride);
     }
 
     public function denyStop(Ride $ride, User $user){
         //dd($ride);
         $ride->users()->detach($user->id);
-        //CORREO
+        $this->sendDenyMail($user, $ride);
         return redirect()->route('ride.show', $ride);
+    }
+
+    public function sendMail(User $user, Ride $ride, Place $place){
+        $mailable = new confimartionMail($ride, $place);
+        Mail::to($user->email)->send($mailable);
+    }
+
+    public function sendDenyMail(User $user, Ride $ride){
+        $mailable = new denyMail($ride);
+        Mail::to($user->email)->send($mailable);
     }
 
 }
